Assignment 1/requirements.txt
absl-py==0.15.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==35.0.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.16.1
flit_core==3.7.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.9.0
ipykernel==5.5.5
ipython==7.33.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.5
jsonschema==4.16.0
jupyter-client==7.0.6
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.6
jupyterlab-pygments==0.2.2
jupyterlab-server==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.2.2
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.3
nbclassic==0.4.3
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.19.2
opt-einsum==3.3.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.1
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==19.0.2
requests==2.28.1
scipy==1.1.0
Send2Trash==1.8.0
setuptools==63.4.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
tornado==6.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==0.16.1
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.0
